pyproject.toml
[tool.poetry]
name = "youtube-python"
version = "0.1.0"
description = ""
authors = ["Sobia Khalid <sobia.khalid@colorifix.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.172.0"
google-genai = "^1.29.0"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
